Add gem stuff & merge with node stuff
In order to have the gem and npm packages sourced from the same
repo, we will have to modify the bin/ directory slightly. This is
because gems require the main executable to be a ruby script. Thus,
we must rename the spritesh shell script to spritesh.sh, and have
a ruby script (spritesh) that executes spritesh.sh.

Since we don't want two copies of the spritesh shell script, I
added a Makefile that copies the script from src/ to the necessary
locations.